requirements.txt
pyserial~=3.5
setuptools~=44.1.1